server/requirements.txt
python-chess>=1.10.0
stockfish>=3.28.0
flask>=2.3.3